fix: consolidate replay metadata reader
